Added hooks for multiple (heterogeneous) microblaze cores support
git-svn-id: file:///dev/shm/git-devel/svn-repo@11238 0848cbd7-7569-0410-ab05-8cc0a0c67708

Former-commit-id: 8aabd7f8a0a5e3df0fd4fe3882c8fe6f19c13053
Former-commit-id: 3c232c82eb426600ab66c6ea36d1ff6076ba3ef5
Former-commit-id: acb6b8946b373eb22fbab2c9bc23c26cbd32fb4e

+++ StdLib/trunk/build/vivado_sdk_bit_v1.tcl
@@ -16,20 +16,27 @@ source -quiet ${VIVADO_BUILD_DIR}/vivado_proc_v1.tcl
 # Open the project (in case not already opened)
 open_project -quiet ${VIVADO_PROJECT}
 
-# Generate .ELF
-set src_rc [catch {exec xsdk -batch -source ${VIVADO_BUILD_DIR}/vivado_sdk_elf_v1.tcl >@stdout}]    
+# Check if custom SDK exist
+if { [file exists ${VIVADO_DIR}/sdk.tcl] == 1 } {   
+   source ${VIVADO_DIR}/sdk.tcl
+} else {
 
-# Add .ELF to the .bit file properties
-add_files -norecurse ${SDK_ELF}  
-set_property SCOPED_TO_REF   [file rootname [file tail ${BD_FILES}]] [get_files ${SDK_ELF} ]
-set_property SCOPED_TO_CELLS { microblaze_0 }                        [get_files ${SDK_ELF} ]
+   # Generate .ELF
+   exec xsdk -batch -source ${VIVADO_BUILD_DIR}/vivado_sdk_elf_v1.tcl >@stdout
 
-# Rebuild the .bit file with the .ELF file include
-reset_run impl_1 -prev_step
-launch_runs -to_step write_bitstream impl_1 >@stdout
-set src_rc [catch { 
-   wait_on_run impl_1 
-} _RESULT]  
+   # Add .ELF to the .bit file properties
+   add_files -norecurse ${SDK_ELF}  
+   set_property SCOPED_TO_REF   [file rootname [file tail ${BD_FILES}]] [get_files ${SDK_ELF} ]
+   set_property SCOPED_TO_CELLS { microblaze_0 }                        [get_files ${SDK_ELF} ]
 
-# Copy over .bit w/ .ELF file to image directory
-exec cp -f ${IMPL_DIR}/${PROJECT}.bit ${IMAGES_DIR}/${PROJECT}_${PRJ_VERSION}.bit
+   # Rebuild the .bit file with the .ELF file include
+   reset_run impl_1 -prev_step
+   launch_runs -to_step write_bitstream impl_1 >@stdout
+   set src_rc [catch { 
+      wait_on_run impl_1 
+   } _RESULT]  
+
+   # Copy over .bit w/ .ELF file to image directory
+   exec cp -f ${IMPL_DIR}/${PROJECT}.bit ${IMAGES_DIR}/${PROJECT}_${PRJ_VERSION}.bit
+   
+}
